+++ plugins/AssocMgr/src/utils.cpp
@@ -112,17 +112,17 @@ BOOL EnumDbPrefixSettings(const char *pszModule,const char *pszSettingPrefix,cha
 {
 	DBCONTACTENUMSETTINGS dbces;
 	struct EnumPrefixSettingsParams param;
-	dbces.szModule=pszModule;
-	dbces.pfnEnumProc=EnumPrefixSettingsProc;
-	dbces.lParam=(LPARAM)&param;
-	param.settings=NULL;
-	param.nSettingsCount=0;
-	param.pszPrefix=pszSettingPrefix;
-	param.nPrefixLen=mir_strlen(pszSettingPrefix);
-	CallService(MS_DB_CONTACT_ENUMSETTINGS,0,(LPARAM)&dbces);
-	*pnSettingsCount=param.nSettingsCount;
-	*pSettings=param.settings;
-	return param.nSettingsCount!=0;
+	dbces.szModule = pszModule;
+	dbces.pfnEnumProc = EnumPrefixSettingsProc;
+	dbces.lParam = (LPARAM)&param;
+	param.settings = NULL;
+	param.nSettingsCount = 0;
+	param.pszPrefix = pszSettingPrefix;
+	param.nPrefixLen = (int)mir_strlen(pszSettingPrefix);
+	CallService(MS_DB_CONTACT_ENUMSETTINGS, 0, (LPARAM)&dbces);
+	*pnSettingsCount = param.nSettingsCount;
+	*pSettings = param.settings;
+	return param.nSettingsCount != 0;
 }
 
 /************************* Error Output ***************************/
